backend/requirements.txt
APScheduler==3.10.4
SQLAlchemy==2.0.34
alembic==1.13.2
bcrypt==3.2.2
docxtpl==0.16.7
email-validator==2.2.0
fastapi==0.115.0
jinja2==3.1.4
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pyjwt==2.8.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
reportlab==4.2.4
slowapi==0.1.9
uvicorn[standard]==0.30.6
